move MainActivity, MainApplication
